requirements.txt
requests
beautifulsoup4
python-dotenv
fake-useragent
flask
